Rename `ast_spec::AstSpec` to `ast::Ast`
