+++ src/ast/mod.rs
@@ -5,7 +5,6 @@ pub mod json;
 pub mod size;
 pub mod test_json;
 
-use crate::node_map::{NodeMap, Reference};
 use display_token::{write_tokens, DisplayToken};
 use size::Size;
 
@@ -14,7 +13,7 @@ use size::Size;
 use crate::editable_tree::EditableTree;
 
 /// The specification of an AST that sapling can edit
-pub trait ASTSpec<Ref: Reference>: std::fmt::Debug + Clone + Eq + Default {
+pub trait Ast: std::fmt::Debug + Clone + Eq + Default {
     /// A type parameter that will represent the different ways this AST can be rendered
     type FormatStyle;
 
@@ -22,30 +21,21 @@ pub trait ASTSpec<Ref: Reference>: std::fmt::Debug + Clone + Eq + Default {
 
     /// Returns an iterator of all the items that need to be rendered to the screen to make up this
     /// node, along with their on-screen locations.
-    fn display_tokens(&self, format_style: &Self::FormatStyle) -> Vec<DisplayToken<Ref>>;
+    fn display_tokens(&self, format_style: &Self::FormatStyle) -> Vec<DisplayToken<Self>>;
 
     /// Determine the space on the screen occupied by this node in an AST
-    fn size(&self, node_map: &impl NodeMap<Ref, Self>, format_style: &Self::FormatStyle) -> Size;
+    fn size(&self, format_style: &Self::FormatStyle) -> Size;
 
     /// Write the textual representation of this AST to a string
-    fn write_text(
-        root: Ref,
-        node_map: &impl NodeMap<Ref, Self>,
-        string: &mut String,
-        format_style: &Self::FormatStyle,
-    ) {
-        write_tokens(root, node_map, string, format_style);
+    fn write_text(&self, string: &mut String, format_style: &Self::FormatStyle) {
+        write_tokens(string, format_style);
     }
 
     /// Make a [`String`] representing this AST.
     /// Same as [`write_text`](ASTSpec::write_text) but creates a new [`String`].
-    fn to_text(
-        root: Ref,
-        node_map: &impl NodeMap<Ref, Self>,
-        format_style: &Self::FormatStyle,
-    ) -> String {
+    fn to_text(&self, format_style: &Self::FormatStyle) -> String {
         let mut s = String::new();
-        Self::write_text(root, node_map, &mut s, format_style);
+        Self::write_text(&mut s, format_style);
         s
     }
 
@@ -53,22 +43,17 @@ pub trait ASTSpec<Ref: Reference>: std::fmt::Debug + Clone + Eq + Default {
 
     /// Get a slice over the direct children of this node.  This operation is expected to be
     /// cheap - it will be used a lot of times without caching the results.
-    fn children(&self) -> &[Ref];
+    fn children(&self) -> &[&Self];
 
     /// Get a mutable slice over the direct children of this node.  Like
     /// [`children`](ASTSpec::children), this operation is expected to be
     /// cheap - it will be used a lot of times without caching the results.
-    fn children_mut(&mut self) -> &mut [Ref];
+    fn children_mut(&mut self) -> &mut [&Self];
 
     /// Get the display name of this node
     fn display_name(&self) -> String;
 
-    fn write_tree_view_recursive(
-        &self,
-        node_map: &impl NodeMap<Ref, Self>,
-        string: &mut String,
-        indentation_string: &mut String,
-    ) {
+    fn write_tree_view_recursive(&self, string: &mut String, indentation_string: &mut String) {
         // Push the node's display name with indentation and a newline
         string.push_str(indentation_string);
         string.push_str(&self.display_name());
@@ -76,10 +61,8 @@ pub trait ASTSpec<Ref: Reference>: std::fmt::Debug + Clone + Eq + Default {
         // Indent by two spaces
         indentation_string.push_str("  ");
         // Write all the children
-        for child_ref in self.children().iter() {
-            if let Some(child) = node_map.get_node(*child_ref) {
-                child.write_tree_view_recursive(node_map, string, indentation_string);
-            }
+        for child in self.children().iter() {
+            child.write_tree_view_recursive(string, indentation_string);
         }
         // Reset indentation
         for _ in 0..2 {
@@ -88,9 +71,9 @@ pub trait ASTSpec<Ref: Reference>: std::fmt::Debug + Clone + Eq + Default {
     }
 
     /// Render a tree view of this node, similar to the output of the Unix command 'tree'
-    fn write_tree_view(&self, node_map: &impl NodeMap<Ref, Self>, string: &mut String) {
+    fn write_tree_view(&self, string: &mut String) {
         let mut indentation_string = String::new();
-        self.write_tree_view_recursive(node_map, string, &mut indentation_string);
+        self.write_tree_view_recursive(string, &mut indentation_string);
         // Pop the unnecessary newline at the end
         let popped_char = string.pop();
         debug_assert_eq!(Some('\n'), popped_char);
@@ -99,9 +82,9 @@ pub trait ASTSpec<Ref: Reference>: std::fmt::Debug + Clone + Eq + Default {
     /// Build a string of the a tree view of this node, similar to the output of the Unix command
     /// 'tree'.  This is the same as [`write_tree_view`](ASTSpec::write_tree_view), except that it
     /// returns a [`String`] rather than appending to an existing [`String`].
-    fn tree_view(&self, node_map: &impl NodeMap<Ref, Self>) -> String {
+    fn tree_view(&self) -> String {
         let mut s = String::new();
-        self.write_tree_view(node_map, &mut s);
+        self.write_tree_view(&mut s);
         s
     }
