+++ src/lib.rs
@@ -1,4 +1,4 @@
 pub mod arena;
-pub mod ast_spec;
+pub mod ast;
 pub mod editable_tree;
 pub mod editor;
